Separating Logic and and CLI
