+++ Epsilon.Cli/Program.cs
@@ -1,4 +1,6 @@
-﻿using Epsilon.Cli;
+﻿using Epsilon;
+using Epsilon.Abstractions;
+using Epsilon.Cli;
 using Epsilon.Extensions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
