+++ Epsilon/Extensions/CoreServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
-﻿using Epsilon.Abstractions.Export;
+﻿using Epsilon.Abstractions;
+using Epsilon.Abstractions.Export;
 using Epsilon.Canvas;
 using Epsilon.Export;
 using Microsoft.Extensions.Configuration;
@@ -21,6 +22,7 @@ public static class CoreServiceCollectionExtensions
         services.Configure<ExportSettings>(config);
         services.AddScoped<ICanvasModuleFileExporter, ConsoleCanvasModuleFileExporter>();
         services.AddScoped<ICanvasModuleFileExporter, CsvCanvasModuleFileExporter>();
+        services.AddScoped<IMainLogic, MainLogic>();
 
         return services;
     }
